refactor(command-visualizer): rewire render/ imports to model/
Repoint render/ (serialize, serialize-cascade, svg-wrap, jumbotron, css,
css-cascade, symbols, html) at ../model/* and ../data/*. serialize.ts now
imports Pos/Range/orderRange from ../model/geometry instead of declaring
them locally. No render/ file imports logic/.

Pure import rewrite; no behavior change.
